+++ src/spinner.js
@@ -1,41 +1,42 @@
-import ora from 'ora'
+import { promisify } from 'util'
 
-export const addSpinner = function({ body }, progress) {
+import ora from 'ora'
+// TODO: replace with `Stream.finished()` after dropping support for Node 8/9
+import endOfStream from 'end-of-stream'
+
+const pEndOfStream = promisify(endOfStream)
+
+// Add CLI spinner showing download progress
+export const addSpinner = async function(response, progress) {
   if (!progress) {
     return
   }
 
-  const spinner = startSpinner(TEXT)
-
-  // eslint-disable-next-line fp/no-let
-  let size = 0
-  body.on('data', buffer => {
-    // eslint-disable-next-line fp/no-mutation
-    size += buffer.length
-    updateSpinner(spinner, size)
-  })
-
-  body.on('end', () => {
-    stopSpinner(spinner)
-  })
-}
-
-const startSpinner = function() {
   const spinner = ora({ color: 'green', spinner: 'star' })
   spinner.start()
-  return spinner
+
+  response.on('downloadProgress', ({ transferred }) => {
+    updateSpinner(spinner, transferred)
+  })
+
+  // TODO: use try/finally after dropping support for Node 8/9
+  try {
+    await pEndOfStream(response, { writable: false })
+    spinner.stop()
+  } catch {
+    // This happens when a network error happens in the middle of the download,
+    // which is hard to simulate in tests
+    // istanbul ignore next
+    spinner.stop()
+  }
 }
 
-const updateSpinner = function(spinner, size) {
+const updateSpinner = function(spinner, transferred) {
   // eslint-disable-next-line fp/no-mutation, no-param-reassign
-  spinner.text = `${TEXT} ${getMegabytes(size)}`
+  spinner.text = `${TEXT} ${getMegabytes(transferred)}`
 }
 
 const TEXT = 'Downloading Node.js...'
-
-const stopSpinner = function(spinner) {
-  spinner.stop()
-}
 
 const getMegabytes = function(size) {
   const sizeA = Math.floor(size / BYTES_TO_MEGABYTES)