+++ test/main.js
@@ -2,36 +2,36 @@ import { env, stderr } from 'process'
 
 import test from 'ava'
 import sinon from 'sinon'
+import getStream from 'get-stream'
 
 import fetchNodeWebsite from '../src/main.js'
 
-const fetchReleases = async function(url = 'index.json', progress = false) {
-  const response = await fetchNodeWebsite(url, { progress })
-  const releases = await response.json()
+const fetchReleases = async function(url, opts) {
+  const response = await fetchNodeWebsite(url, opts)
+  const content = await getStream(response)
+  const releases = JSON.parse(content)
   return releases
 }
 
 test('Error request', async t => {
-  await t.throwsAsync(fetchNodeWebsite('\u0000', { progress: false }))
+  await t.throwsAsync(fetchReleases('\u0000', { progress: false }))
 })
 
 test('Error response', async t => {
   // This release does not exist. We make a request that looks legit so we are
   // not blocked.
   await t.throwsAsync(
-    fetchNodeWebsite('v12.7.1/node-v12.7.1-linux-x64.tar.gz', {
-      progress: false,
-    }),
+    fetchReleases('v12.7.1/node-v12.7.1-linux-x64.tar.gz', { progress: false }),
   )
 })
 
 test('Success', async t => {
-  const releases = await fetchReleases()
+  const releases = await fetchReleases('index.json', { progress: false })
   t.true(Array.isArray(releases))
 })
 
 test('Leading slashes', async t => {
-  const releases = await fetchReleases('/index.json')
+  const releases = await fetchReleases('/index.json', { progress: false })
   t.true(Array.isArray(releases))
 })
 
@@ -39,7 +39,7 @@ test.serial('Empty mirror website', async t => {
   // eslint-disable-next-line fp/no-mutation
   env.NODE_MIRROR = ''
 
-  const releases = await fetchReleases()
+  const releases = await fetchReleases('index.json', { progress: false })
   t.true(Array.isArray(releases))
 
   // eslint-disable-next-line fp/no-delete
@@ -50,7 +50,7 @@ test.serial('Mirror website', async t => {
   // eslint-disable-next-line fp/no-mutation
   env.NODE_MIRROR = 'https://npm.taobao.org/mirrors/node'
 
-  const releases = await fetchReleases()
+  const releases = await fetchReleases('index.json', { progress: false })
   t.true(Array.isArray(releases))
 
   // eslint-disable-next-line fp/no-delete
@@ -60,8 +60,7 @@ test.serial('Mirror website', async t => {
 test.serial('Do not show spinner if opts.progress false', async t => {
   const spy = sinon.spy(stderr, 'write')
 
-  const response = await fetchNodeWebsite('index.json', { progress: false })
-  await response.json()
+  await fetchReleases('index.json', { progress: false })
 
   t.true(spy.notCalled)
 
@@ -71,8 +70,7 @@ test.serial('Do not show spinner if opts.progress false', async t => {
 test.serial('Show spinner if opts.progress true', async t => {
   const spy = sinon.spy(stderr, 'write')
 
-  const response = await fetchNodeWebsite('index.json', { progress: true })
-  await response.json()
+  await fetchReleases('index.json', { progress: true })
 
   t.true(spy.called)
 
@@ -82,8 +80,7 @@ test.serial('Show spinner if opts.progress true', async t => {
 test.serial('Show spinner by default', async t => {
   const spy = sinon.spy(stderr, 'write')
 
-  const response = await fetchNodeWebsite('index.json')
-  await response.json()
+  await fetchReleases('index.json')
 
   t.true(spy.called)